[1.2.0]
### Added

- Update for PHP 8.2 compatibility.

### Tested

- Tested on WooCommerce 9.9 and WordPress 6.8.

### Updated

- Code quality and security fixes.
