+++ CHANGELOG.md
@@ -1,0 +1,72 @@
+# Changelog
+
+## [1.2.0](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.2.0)
+
+### Added
+
+- Update for PHP 8.2 compatibility.
+
+### Tested
+
+- Tested on WooCommerce 9.9 and WordPress 6.8.
+
+### Updated
+
+- Code quality and security fixes.
+
+## [1.1.0](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.1.0)
+
+### Added
+
+- WooCommerce Blocks support.
+- WooCommerce HPOS compatibility.
+- Code quality improvements.
+
+### Tested
+
+- Tested with WooCommerce 9.1.4.
+
+## [1.0.4](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.0.4)
+
+### Changed
+
+- Replaced deprecated order status method.
+
+## [1.0.3](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.0.3)
+
+### Changed
+
+- Updated SiD API call.
+
+### Tested
+
+- Tested with WooCommerce 7.9.0.
+
+## [1.0.2](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.0.2)
+
+### Added
+
+- Buyer Return and Notify URL configuration instructions.
+- Support for PHP 8.0.
+
+### Tested
+
+- Tested with WooCommerce 7.7.2 and WordPress 6.2.2.
+
+## [1.0.1](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.0.1)
+
+### Added
+
+- GitHub updater functionality.
+- SiD Secure EFT branding update.
+
+### Tested
+
+- Tested with WooCommerce 6.0 and WordPress 5.9.
+
+## [1.0.0](https://github.com/SiD-Secure-EFT/SiD_WooCommerce/releases/tag/v1.0.0)
+
+### Added
+
+- Initial release with minor changes and formatting.
+- Updated logo and text.
