+++ readme.txt
@@ -2,8 +2,45 @@
 Contributors: appinlet
 Tags: ecommerce, woocommerce, payment, SiD Payment, credit card
 Requires at least: 6.0
-Tested up to: 6.6.0
+Tested up to: 6.8
 Requires PHP: 8.0
-Stable tag: 1.1.0
+Stable tag: 1.2.0
 License: GPLv3
 License URI: https://www.gnu.org/licenses/gpl-3.0.html
+
+== Changelog ==
+
+= 1.2.0 - 2025-07-27 =
+ * Update for PHP 8.2 compatibility.
+ * Tested on WooCommerce 9.9 and WordPress 6.8.
+ * Code quality and security fixes.
+
+= 1.1.0 - 2024-08-16 =
+ * Added WooCommerce Blocks support.
+ * Tested with WooCommerce 9.1.4.
+ * Code quality improvements.
+ * Added WooCommerce HPOS compatibility.
+
+= 1.0.4 - 2023-07-31 =
+ * Replaced deprecated order status method.
+
+= 1.0.3 - 2023-07-26 =
+ * Tested with WooCommerce 7.9.0.
+ * Updated SiD API call.
+
+= 1.0.2 - 2023-06-07 =
+ * Tested with WooCommerce 7.7.2 and WordPress 6.2.2.
+ * Updated for PHP 8.0 compatibility.
+ * Added Buyer Return and Notify URL configuration instructions.
+
+= 1.0.1 - 2022-02-03 =
+ * Added GitHub updater functionality.
+ * Tested with WooCommerce 6.0 and WordPress 5.9.
+ * Updated SiD Secure EFT branding.
+
+= 1.0.0 - 2020-02-14 =
+ * Initial release with minor changes and formatting.
+ * Updated logo and text.
+
+[See changelog for all versions](https://raw.githubusercontent.com/SiD-Secure-EFT/SiD_WooCommerce/master/CHANGELOG.md).
+
